Add english lang
Fri 3/6/2020 8:57 AM
I'm sorry to bother you with this again Gary has requested another field to be added to the Helios 1 report we were working from. Can you please go into the report and add English as 1st language or primary language. Which ever will indicate those who English is the primary language and  those for which English is not.

Need today ASAP. Sorry to spring it on you.